Update AdvertiserContentArticle.json
For Advertiser Content, The Trust Project based on testing with platforms has asked newsrooms to attach only the "AdvertiserContentArticle" markup. All the other TTP specified markup such as NewsMediaOrganization, citations, corrections, author attributes etc, need not be marked up on this type of page. 

For e.g. on this JSON page, I've removed all T-related markup except the ToW tag. I left all the other metadata that is part of general schema.org markup for articles intact. 

We may do one more iteration of this in the next week or so, so please do not commit to your IT team as approved page schema for Ad Content Articles yet!